backend/requirements.txt
Werkzeug==0.16.1
Flask==0.12.2
google-api-python-client==2.28.0
google-cloud-storage==1.42.3